pyproject.toml
[tool.poetry]
name = "unreverso"
version = "0.1.0"
description = ""
authors = ["nik <i@nikmosi.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
epitran = "^1.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
